Move out dot, mat_mul
